LayoutTests/http/tests/cache/disk-cache/disk-cache-fragments.html

<script src="/js-test-resources/js-test-pre.js"></script>
<script src="resources/cache-test.js"></script>
<body>
<script>

var testMatrix =
[
 [
  { responseHeaders: { 'Cache-control': 'max-age=100' } },
  ],
];

description("Test that URLs with different fragments should be cached.");

var tests = generateTests(testMatrix);
debug("running " + tests.length + " tests");
debug("");

requireFragment(tests);
runTests(tests);

</script>
<script src="/js-test-resources/js-test-post.js"></script>
