LayoutTests/http/tests/cache/disk-cache/disk-cache-fragments-expected.txt

Test that URLs with different fragments should be cached.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


running 1 tests

--------Testing loads from disk cache--------
response headers: {"Cache-control":"max-age=100"}
response source: Disk cache

--------Testing loads through memory cache (XHR behavior)--------
response headers: {"Cache-control":"max-age=100"}
response source: Memory cache

--------Testing loads through memory cache (subresource behavior)--------
response headers: {"Cache-control":"max-age=100"}
response source: Memory cache

PASS successfullyParsed is true

TEST COMPLETE
